rename OpenSSL_Client sub project as client
