+++ CMakeLists.txt
@@ -8,4 +8,4 @@ project ("OpenSSL_101")
 find_package(OpenSSL)
 
 # Include sub-projects.
-add_subdirectory ("OpenSSL_Client")
+add_subdirectory ("client")
