--- OpenSSL_Client/OpenSSL_Client.h
@@ -1,8 +1,0 @@
-﻿// OpenSSL_Client.h : Include file for standard system include files,
-// or project specific include files.
-
-#pragma once
-
-#include <iostream>
-
-// TODO: Reference additional headers your program requires here.
