+++ client/client.h
@@ -1,0 +1,8 @@
+﻿// OpenSSL_Client.h : Include file for standard system include files,
+// or project specific include files.
+
+#pragma once
+
+#include <iostream>
+
+// TODO: Reference additional headers your program requires here.
